parallel try but fail

+++ src/model/IsingModel.cpp
@@ -39,27 +39,21 @@ void IsingModel::simul(Grid &grid) {
     //boost::uniform_real<> uniformReal(0.0, 1.0);
     //boost::variate_generator<RNGType, boost::uniform_real<> > dice(rng, uniformReal);
     double localEnergy = 0;
-    double currentSpin = 0;
+    int currentSpin = 0;
     double alea = 0;
     tempoGrid = grid;
     for (int i = 0; i < size; i++) {
         for (int j = 0; j < size; j++) {
             localEnergy = grid.getLocalEnergy(i, j) - grid.getTotalSpin() * B;
-            currentSpin = grid.getMatrix()->get(i, j);
-            //cout << currentSpin << endl;
+            currentSpin = (int) grid.getMatrix()->get(i, j);
             if (localEnergy < 0) {
                 tempoGrid.getMatrix()->set(i, j, -currentSpin);
-                //cout << "inversion energie négative" << endl;
             } else {
                 alea = distribution(generator);
-                //cout << "inversion alea : " << alea << " | compare to :" << exp(-localEnergy / (k * T)) << endl;
-                //tempoGrid.getMatrix()->set(i, j, currentSpin);
                 if (alea < exp(-localEnergy / (k * T))) {
                     tempoGrid.getMatrix()->set(i, j, -currentSpin);
-                    //cout << "inversion alea" << endl;
                 }
             }
-            //cout << tempoGrid.getMatrix()->get(i,j) << endl;
         }
     }
     grid = tempoGrid;
